# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API behavior, compatibility rules, and endpoint contract requirements.",
    "version": "1.25.1"
  },
  "paths": {
...
    "/api/wip/overview/summary": {
      "get": {
        "summary": "GET /api/wip/overview/summary",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/500"
      },
      "post": {
        "summary": "POST /api/wip/overview/summary",
        "responses": {
          "201": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Contract request: JSON body (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "400/500"
      }
    },
    "/api/wip/overview/matrix": {
      "post": {
        "summary": "POST /api/wip/overview/matrix",
        "responses": {
          "201": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Contract request: JSON body (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "400/500"
      },
      "get": {
        "summary": "GET /api/wip/overview/matrix",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/500"
      }
    },
    "/api/wip/overview/hold": {
      "get": {
        "summary": "GET /api/wip/overview/hold",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/500"
      },
      "post": {
        "summary": "POST /api/wip/overview/hold",
        "responses": {
          "201": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Contract request: JSON body (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "400/500"
      }
    },
    "/api/wip/detail/{workcenter}": {
      "get": {
        "summary": "GET /api/wip/detail/{workcenter}",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "parameters": [
          {
            "name": "workcenter",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/500"
      },
      "post": {
        "summary": "POST /api/wip/detail/{workcenter}",
        "responses": {
          "201": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "parameters": [
          {
            "name": "workcenter",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Contract request: JSON body (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "400/500"
      }
    },
    "/api/wip/lot/{lotid}": {
      "get": {
        "summary": "GET /api/wip/lot/{lotid}",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "404": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "parameters": [
          {
            "name": "lotid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "404/500"
      }
    },
    "/api/wip/meta/workcenters": {
      "get": {
        "summary": "GET /api/wip/meta/workcenters",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "500"
      }
    },
    "/api/wip/meta/packages": {
      "get": {
        "summary": "GET /api/wip/meta/packages",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "500"
      }
    },
    "/api/wip/meta/filter-options": {
      "get": {
        "summary": "GET /api/wip/meta/filter-options",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "500"
      },
      "post": {
        "summary": "POST /api/wip/meta/filter-options",
        "responses": {
          "201": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Contract request: JSON body (schema not machine-resolved; see contract)",
          "content": {},
          "x-cdd-unresolved": true
        },
        "x-cdd-errors": "500"
      }
    },
    "/api/wip/meta/search": {
      "get": {
        "summary": "GET /api/wip/meta/search",
        "responses": {
          "200": {
            "description": "Contract response: GenericSuccessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error response (see contract error format)"
          },
          "500": {
            "description": "Error response (see contract error format)"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-cdd-errors": "400/500"
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "GenericSuccessResponse": {
        "type": [
          "object",
          "null"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {},
          "error": {},
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string"
              },
              "app_version": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}